requirements.txt
wget~=3.2
requests~=2.25.1
pyttsx3~=2.90
future~=0.18.2
PyGithub~=1.55
Pillow~=8.2.0